+++ src/main/groovy/io/seqera/wave/service/k8s/K8sServiceImpl.groovy
@@ -443,7 +443,7 @@ class K8sServiceImpl implements K8sService {
         }
     }
 
-/**
+    /**
      * Fetch the logs of a pod
      *
      * @param name The pod name
