+++ src/main/groovy/io/seqera/wave/service/packages/DockerPackagesService.groovy
@@ -34,7 +34,7 @@ import jakarta.inject.Singleton
 @Singleton
 @Requires(missingProperty = 'wave.build.k8s')
 @CompileStatic
-class DockerCondaFetcher extends AbstractCondaFetcher {
+class DockerPackagesService extends AbstractPackagesService {
 
     @Inject
     PackagesConfig config
